Box::effectiveWidth returned int for a short claim, so MSVC refused the Toolbox layout assignment (C4244): return the claim's type
The width claim (`BoxProps::width`) and every consumer (`LayoutState::width`, the Row seat claim, the Toolbox handlers) are `short`; the State door is `State<int>` for the application's convenience, so the narrowing belongs in the one accessor, not at eleven call sites.

+++ common/app/nodes/nestable/Box.hpp
@@ -53,9 +53,9 @@ namespace loka
         return this->effectiveWidth() > 0 && height > 0;
       }
       /** Resolves the width used by Box layout and its parent's seat claim. */
-      int effectiveWidth() const
+      short effectiveWidth() const
       {
-        return this->widthState_ ? this->widthState_->get() : this->width;
+        return this->widthState_ ? static_cast<short>(this->widthState_->get()) : this->width;
       }
       bool operator<(const scene::PropsBase &rhs) const
       {
